한줄평 체크 js 수정
