+++ src/main/resources/static/css/book.css
@@ -22,6 +22,7 @@
 }
 .book_side_bar li:hover {
   background-color: #f0f0f0;
+  transition: .3s;
 }
 
 /*search*/
@@ -255,6 +256,18 @@
 }
 
 /*별점*/
+.starAvgText {
+  color: rgb(255, 99, 132);
+  font-weight: 600;
+  border-radius: 10px;
+  padding: 5px;
+}
+.starAvgText:hover {
+  background-color:rgba(255, 222, 89);
+  color: #000;
+  transition: .2s;
+  
+}
 .detail_score {
   display: flex;
 }
@@ -341,7 +354,7 @@
 }
 .el_good {
   display: inline-block;
-  background-color: rgba(255, 222, 89,.8);
+  background-color: #f0f0f0;
   padding: 2px 4px;
   border-radius: 8px;
   margin-right: 10px;
