+++ src/main/resources/static/js/book.js
@@ -42,8 +42,8 @@ function saveReview() {
     return false; // 이벤트 막기
    }
    
-   const checkWish = document.getElementById('checkWish').value;
-   if (checkWish !== 0) {
+   const checkScore = document.getElementById('checkScore').value;
+   if (checkScore != 0) {
      alert("이미 한줄평을 등록했습니다. 도서 당 하나만 등록 가능합니다");
      return false; // 이벤트 막기
    }
@@ -235,3 +235,21 @@ $(document).ready(function() {
   })
 });
 
+//그래프
+
+document.addEventListener('DOMContentLoaded', function () {
+    var ctx = document.getElementById('barChart').getContext('2d');
+    
+    var chart = new Chart(ctx, {
+        type: 'bar', // 
+        data: {
+            labels: ['1점', '2점', '3점', '4점','5점'],
+            datasets: [{
+                label: '평점 분포',
+                backgroundColor: 'rgb(255, 99, 132)',
+                borderColor: 'rgb(255, 99, 132)',
+                data: [2, 10, 5,3,3]
+            }]
+        },
+    });
+});
