+++ src/main/resources/templates/book/result.html
@@ -54,7 +54,7 @@
           <ul id="book_side_list">
             <li><a th:href="@{/mypage/booklist.do}">자료검색</a></li>
             <li><a th:href="@{/mypage/edit.form}">주제별자료검색</a></li>
-            <li><a th:href="@{/mypage/wish.do}">희망도서신청</a></li>
+            <li><a th:href="@{/apply/bookapply.do}">희망도서신청</a></li>
             <li><a th:href="@{/mypage/review.do}">사서추천도서</a></li>
           </ul>      
         </div>
